소스코드/장고/wherego/requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.12.1
Automat==22.10.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
constantly==15.1.0
contourpy==1.0.6
cryptography==37.0.4
cycler==0.11.0
daphne==4.0.0
defusedxml==0.7.1
Django==3.2
django-allauth==0.51.0
django-session-timeout==0.1.0
djangorestframework==3.14.0
filelock==3.9.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.0
h11==0.13.0
h5py==3.7.0
haversine==2.7.0
huggingface-hub==0.11.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==5.0.0
incremental==22.10.0
joblib==1.2.0
JPype1==1.4.1
keras==2.10.0
Keras-Preprocessing==1.1.2
keybert==0.7.0
kiwisolver==1.4.4
konlpy==0.6.0
krwordrank==1.0.3
libclang==14.0.6
lightgbm==3.3.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
msgpack==1.0.4
nltk==3.8
numpy==1.23.1
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.4.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
selenium==4.3.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.19.0
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow_intel==2.10.0
termcolor==2.1.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.25.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.11
Werkzeug==2.2.2
wordcloud==1.8.2.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.10.0
zope.interface==5.5.2
